Add: authorization

+++ src/components/auth/Registration.jsx
@@ -1,6 +1,8 @@
 import {useState} from "react";
 import {NavLink} from "react-router-dom";
 import '../../style/auth/Registration.css'
+import Modal from "react-bootstrap/Modal";
+import Button from "react-bootstrap/Button";
 
 const Registration = () => {
     const [lastname, setLastname] = useState("");
@@ -9,6 +11,7 @@ const Registration = () => {
     const [password, setPassword] = useState("");
     const [confirmedPassword, setConfirmedPassword] = useState("");
     const [isModal, setIsModal] = useState(false);
+    const [errorMessage, setErrorMessage] = useState("")
 
     function sendRequest() {
         const reqBody = {
@@ -18,36 +21,41 @@ const Registration = () => {
             hashPassword: password
         };
 
-        let cache = [];
-        fetch("http://localhost:8080/auth/register", {
-            method: 'POST',
-            headers: {
-                'Content-Type': 'application/json'
-            },
-            body: JSON.stringify(reqBody, function(key, value) {
-                if (typeof value === 'object' && value !== null) {
-                    if (cache.indexOf(value) !== -1) {
-                        return;
-                    }
-                    cache.push(value);
-                }
-                return value;
-            })
-        }).then((response) => {
-            if (response.status === 200) {
-                return Promise.all([response.json()]);
-            } else {
-                return Promise.reject("Invalid data.");
-            }
-        })
-            .then(data => {
-                localStorage.setItem("jwtToken", data[0].jwtToken);
-                localStorage.setItem("currentUserRole", data[0].userRole);
-                localStorage.setItem("currentUserId", data[0].userId);
-                window.location.assign("/profile");
-            }).catch(() => {
+        if (confirmedPassword !== password) {
+            setErrorMessage("The passwords don't match.")
             setIsModal(true);
-        });
+        } else {
+            let cache = [];
+            fetch("http://localhost:8080/auth/register", {
+                method: 'POST',
+                headers: {
+                    'Content-Type': 'application/json'
+                },
+                body: JSON.stringify(reqBody, function (key, value) {
+                    if (typeof value === 'object' && value !== null) {
+                        if (cache.indexOf(value) !== -1) {
+                            return;
+                        }
+                        cache.push(value);
+                    }
+                    return value;
+                })
+            }).then((response) => {
+                if (response.status === 200) {
+                    return Promise.all([response.json()]);
+                } else {
+                    return Promise.reject("Invalid data.");
+                }
+            })
+                .then(data => {
+                    localStorage.setItem("jwtToken", data[0].jwtToken);
+                    localStorage.setItem("currentUserRole", data[0].userRole);
+                    localStorage.setItem("currentUserId", data[0].userId);
+                    window.location.assign("/profile");
+                }).catch(() => {
+                setIsModal(true);
+            });
+        }
     }
 
     return (
@@ -81,6 +89,17 @@ const Registration = () => {
                 <button id={"registrationRegisterButton"} type="button" className="btn btn-dark"
                         onClick={() => sendRequest()}>REGISTER</button>
             </div>
+            <Modal show={isModal} onHide={() => setIsModal(false)}>
+                <Modal.Header closeButton>
+                    <Modal.Title>Warning!</Modal.Title>
+                </Modal.Header>
+                <Modal.Body>{errorMessage}</Modal.Body>
+                <Modal.Footer>
+                    <Button variant="secondary" onClick={() => setIsModal(false)}>
+                        Close
+                    </Button>
+                </Modal.Footer>
+            </Modal>
         </div>
     )
 }
